Fix Azure MB alignment issue
Previously, qemu-img would have extended new image virtual sizes
based on the CHS algorithm provided by the VHD specs in order
to ensure that the disk geometry (and payload as seen by some
guests which use the CHS value) can fit in the requested disk.

This patch drops this behavior, as it breaks compatibility with
Azure, which requires the MB alignment to be preserved.